dev: bot.py cog and started twitch_notifications cog
